Use case added UpdateSaleStatusUseCase, RemoveFromStockUseCase to keep full track for stock when sale is cancelled, Other changes includes int to Long data type for product Id and using Assisted Factory in viewmodels

+++ app/src/main/java/com/imecatro/demosales/navigation/products/ProductsDestinations.kt
@@ -10,9 +10,9 @@ sealed class ProductsDestinations {
     object Add
 
     @Serializable
-    data class Edit(val id: Int)
+    data class Edit(val id: Long)
 
     @Serializable
-    data class Details(val id: Int)
+    data class Details(val id: Long)
 }
