+++ demosales-ui/sales/src/main/java/com/imecatro/demosales/ui/sales/add/model/ProductResultUiModel.kt
@@ -3,7 +3,7 @@ package com.imecatro.demosales.ui.sales.add.model
 import android.net.Uri
 
 data class ProductResultUiModel(
-    val id: Int = 0,
+    val id: Long = 0,
     val name: String = "",
     val price: Double = 0.0,
     val imageUri: Uri? = null
